requirements.txt
pytesseract 
Pillow
opencv

